+++ src/utils/balancer/contentHash.ts
@@ -5,7 +5,7 @@ import { Provider } from '@ethersproject/abstract-provider';
 import { namehash } from '@ethersproject/hash';
 import { isHexString } from '@ethersproject/bytes';
 import bs58 from 'bs58';
-import { call } from '@/utils/balancer/multicall';
+import { call } from '@/utils/balancer/contract';
 const supportedCodecs = ['ipns-ns', 'ipfs-ns', 'swarm-ns', 'onion', 'onion3'];
 
 const REGISTRAR_ABI = [
