+++ src/utils/balancer/vault.ts
@@ -1,5 +1,5 @@
 import { JsonRpcProvider, Web3Provider } from '@ethersproject/providers';
-import { Multicaller } from '@/utils/balancer/multicall';
+import { Multicaller } from '@/utils/balancer/contract';
 import { call, sendTransaction } from '@/utils/balancer/web3';
 import constants from './constants';
 import { abi } from './abi/Vault.json';
